🎨 Update project structure
