+++ ACKategories-iOSTests/Extensions/UIWindow.swift
@@ -1,0 +1,14 @@
+//
+//  UIWindow.swift
+//  ACKategories-iOSTests
+//
+//  Created by Lukáš Hromadník on 23.11.2020.
+//
+
+import UIKit
+
+extension UIWindow {
+    static var dummy: UIWindow {
+        UIWindow(frame: UIScreen.main.bounds)
+    }
+}
